Move create_filename() into utils
