+++ src/common/util/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(UTIL_SOURCES
   util/demangle.cpp
+  util/file_util.cpp
   util/format_util.cpp
   util/parse_util.cpp
   util/unitfmt.c
