+++ src/services/recorder/Recorder.cpp
@@ -16,16 +16,9 @@
 
 #include "caliper/reader/CaliWriter.h"
 
-#ifdef _WIN32
-#include <process.h>
-#else
-#include <unistd.h>
-#endif
+#include "../../common/util/file_util.h"
 
-#include <algorithm>
-#include <chrono>
 #include <iostream>
-#include <random>
 #include <string>
 
 using namespace cali;
@@ -50,36 +43,6 @@ const ConfigSet::Entry configdata[] = {
     ConfigSet::Terminator
 };
 
-// --- helpers
-
-string random_string(string::size_type len)
-{
-    static std::mt19937 rgen(chrono::system_clock::now().time_since_epoch().count());
-
-    static const char characters[] = "0123456789"
-        "abcdefghiyklmnopqrstuvwxyz"
-        "ABCDEFGHIJKLMNOPQRSTUVWXZY";
-
-    std::uniform_int_distribution<> random(0, sizeof(characters)-2);
-
-    string s(len, '-');
-
-    generate_n(s.begin(), len, [&](){ return characters[random(rgen)]; });
-
-    return s;
-}
-
-string create_filename()
-{
-    char   timestring[16];
-    time_t tm = chrono::system_clock::to_time_t(chrono::system_clock::now());
-    strftime(timestring, sizeof(timestring), "%y%m%d-%H%M%S", localtime(&tm));
-
-    int  pid = static_cast<int>(getpid());
-
-    return string(timestring) + "_" + std::to_string(pid) + "_" + random_string(12) + ".cali";
-}
-
 void write_output_cb(Caliper* c, Channel* chn, const SnapshotRecord* flush_info)
 {
     ConfigSet cfg = chn->config().init("recorder", configdata);
@@ -88,7 +51,7 @@ void write_output_cb(Caliper* c, Channel* chn, const SnapshotRecord* flush_info)
     std::string directory = cfg.get("directory").to_string();
 
     if (filename.empty())
-        filename = create_filename();
+        filename = cali::util::create_filename();
     if (!directory.empty())
         filename = directory + "/" + filename;
